// dunnit-web/src/api/schema.d.ts
export interface components {
    schemas: {
        Auth0Config: {
            domain: string;
            clientId: string;
            audience: string;
        };
        ConfigResponse: {
            testValue: null | string;
            env: null | string;
            auth0: components["schemas"]["Auth0Config"];
        };
        CreateItemRequest: {
            title: string;
            description: null | string;
        };
        CreateListRequest: {
            name: string;
        };
        ItemResponse: {
            /** Format: uuid */
            id: string;
            /** Format: uuid */
            listId: string;
            title: string;
            description: null | string;
            /** Format: double */
            sortOrder: number | string;
            completed: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        ListResponse: {
            /** Format: uuid */
            id: string;
            /** Format: uuid */
            creatorUserId: string;
            name: string;
            /** Format: double */
            sortOrder: number | string;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
        UpdateItemPositionRequest: {
            /** Format: uuid */
            precedingItemId: null | string;
        };
        UpdateItemRequest: {
            title: string;
            description: null | string;
            completed: boolean;
        };
        UpdateListPositionRequest: {
            /** Format: uuid */
            precedingListId: null | string;
        };
        UpdateListRequest: {
            name: string;
        };
        UserResponse: {
            /** Format: uuid */
            id: string;
            email: string;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}